+++ InscryptionAPI/Costs/CardCostManager.cs
@@ -524,6 +524,9 @@ public static class CardCostManager
         while (list.Count > 3)
             list.RemoveAt(SeededRandom.Range(0, list.Count, randomSeed++));
 
+        if (InscryptionAPIPlugin.configRandomChoiceOrder.Value)
+            list = list.Randomize().ToList();
+
         __result = list;
 
     }