+++ InscryptionAPI/InscryptionAPIPlugin.cs
@@ -33,6 +33,7 @@ public class InscryptionAPIPlugin : BaseUnityPlugin
     public static string Directory = "";
 
     internal static ConfigEntry<bool> configOverrideArrows;
+    internal static ConfigEntry<bool> configRandomChoiceOrder;
     internal static ConfigEntry<bool> configHideAct1BossScenery;
     internal static ConfigEntry<TotemManager.TotemTopState> configCustomTotemTopTypes;
     internal static ConfigEntry<ConsumableItemManager.ConsumableState> configCustomItemTypes;
@@ -106,6 +107,7 @@ public class InscryptionAPIPlugin : BaseUnityPlugin
         configCustomTotemTopTypes = Config.Bind("Totems", "Top Types", TotemManager.TotemTopState.CustomTribes, "If Vanilla, don't change totem tops; if CustomTribes, added custom tribes will use custom totem tops; if AllTribes then all totem tops will use a custom top.");
         configCustomItemTypes = Config.Bind("Items", "Types", ConsumableItemManager.ConsumableState.Custom, "If Vanilla, only vanilla items will be used; if Custom, added custom items will use custom models; if All then all items will use a custom model.");
         configOverrideArrows = Config.Bind("Menus", "Override Arrows", false, "When true, forces the challenge screen arrows to appear at the top of the screen instead of the sides.");
+        configRandomChoiceOrder = Config.Bind("Miscellaneous", "Randomise Cost Choice Order", false, "When true, randomises the order card cost choices are presented in Act 1.");
         configHideAct1BossScenery = Config.Bind("Optimization", "Hide Act 1 Scenery", false, "When true bosses will not spawn their scenery. (eg: Prospector's trees) This can improve performance on low-end machines.");
     }